+++ cypress/integration/main/Encounters.spec.ts
@@ -53,7 +53,15 @@ describe('Encounters', () => {
     cy.get('[data-testid=edit-encounter-0]').click();
     cy.contains('Cancel').click();
     cy.get('[data-testid=status-0]').click();
-    cy.get('[data-testid=status-0] > .visible > :nth-child(2)').click();
+    cy.get('[data-testid="status-0"] > .visible > :nth-child(2)').click({
+      waitForAnimations: true,
+      force: true,
+    });
+    cy.get('[data-testid=status-0]').click();
+    cy.get('[data-testid="status-0"] > .visible > :nth-child(2)').click({
+      waitForAnimations: true,
+      force: true,
+    });
     cy.get('[data-testid=status-0] > .divider').should('have.text', 'Fainted');
     cy.get('[data-testid=edit-encounter-0]').click();
     cy.get('[data-testid="cause of fainting"]')