+++ src/common/PokemonSelector/PokemonSelector.tsx
@@ -128,6 +128,7 @@ function PokemonSelector({
     >
       <Modal.Content className={styles.content} scrolling>
         <Filter hideGen={!!limitGen} values={values} />
+        {/* @ts-ignore */}
         <FixedSizeList
           height={400}
           itemCount={filteredPokemon.length}